// src/types.ts
export interface User {
    readonly id: number;
    readonly name: string;
    readonly username: string;
    readonly email: string;
    readonly address: {
        readonly street: string;
        readonly suite: string;
        readonly city: string;
        readonly zipcode: string;
        readonly geo: {
            readonly lat: string;
            readonly lng: string;
        };
    };
    readonly phone: string;
    readonly website: string;
    readonly company: {
        readonly name: string;
        readonly catchPhrase: string;
        readonly bs: string;
    };
}

export interface Todo {
    readonly id: number;
    readonly userId: number;
    readonly title: string;
    readonly completed: boolean;
}